undo changes to config and reshape text array
